% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwise_pool.R
\name{colwise.pool}
\alias{colwise.pool}
\title{Title}
\usage{
colwise.pool(data, n1, n2, m1, m2, sd1, sd2)
}
\arguments{
\item{data}{Data frame of d_pooled_sd output for multiple samples.}

\item{n1}{Sample sizes of group1.}

\item{n2}{Sample sizes of group2.}

\item{m1}{Means of group1.}

\item{m2}{Means of group2.}

\item{sd1}{Standard deviations of group1.}

\item{sd2}{Standard deviations of group2.}
}
\value{
pooled SDs for groups and across groups
}
\description{
Title
}
