% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvv.R
\name{cvv}
\alias{cvv}
\title{Coefficient of variance variation}
\usage{
cvv(data)
}
\arguments{
\item{data}{Data frame of two or more columns or list of two or more variables.}
}
\value{
A vector including VR, CVV, and SVH.
}
\description{
Calculates three different indices for variation between two or more variance estimates.
VR = Variance ratio between the largest and the smallest variance.
CVV = Coefficient of variance variation (Box, 1954).
SVH = Standardized variance heterogeneity (Ruscio & Roche, 2012).
}
\examples{
d <- list(
  X1 = rnorm(10, sd = 10),
  X2 = rnorm(100, sd = 7.34),
  X3 = rnorm(1000, sd = 6.02),
  X4 = rnorm(100, sd = 5.17),
  X5 = rnorm(10, sd = 4.56)
)
cvv(d)
}
\references{
Box, G. E. P. (1954). Some Theorems on Quadratic Forms Applied in the Study of Analysis of Variance Problems, I. Effect of Inequality of Variance in the One-Way Classification. The Annals of Mathematical Statistics, 25(2), 290–302.

Ruscio, J., & Roche, B. (2012). Variance Heterogeneity in Published Psychological Research: A Review and a New Index. Methodology, 8(1), 1–11. https://doi.org/10.1027/1614-2241/a000034
}
