% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_regularized.R
\name{D_regularized}
\alias{D_regularized}
\title{Title}
\usage{
D_regularized(
  data,
  mv.vars,
  group.var,
  group.values,
  alpha = 0.5,
  nfolds = 10,
  s = "lambda.min",
  type.measure = "deviance",
  rename.output = T
)
}
\arguments{
\item{data}{A data.frame}

\item{mv.vars}{Character vector. Variable names in the multivariate variable set}

\item{group.var}{The name of the group variable}

\item{group.values}{Vector of length 2, group values (e.g. c("male", "female) or c(0,1))}

\item{alpha}{Alpha-value for penalizing function ranging from 0 to 1: 0 = ridge regression, 1 = lasso, 0.5 = elastic net (default)}

\item{nfolds}{Number of folds used for obtaining lambda (range from 3 to n-1, default 10)}

\item{s}{Which lambda value is used for predicted values? Either "lambda.min" (default) or "lambda.1se"}

\item{type.measure}{Which measure is used during cross-validation. Default "deviance"}

\item{rename.output}{Logical. Should the output values be renamed according to the group.values? Default TRUE.}
}
\value{

}
\description{
Title
}
