% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_dv_sem.R
\name{diff_dv_sem}
\alias{diff_dv_sem}
\title{Algebraic difference scores as dependent variables in regression analysis conducted via structural equation modeling}
\usage{
diff_dv_sem(
  data,
  var1,
  var2,
  predictor,
  estimator = "MLR",
  sampling.weights = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{var1}{Character string. Variable name of first component of difference score (Y_1).}

\item{var2}{Character string. Variable name of second component of difference score (Y_2).}

\item{predictor}{Character string. Variable name of independent variable predicting difference.}

\item{estimator}{Character string. Estimator used in SEM (Default "MLR").}

\item{sampling.weights}{Character string. Name of sampling weights variable.}
}
\value{
\item{aic}{Fit statistics and model probabilities for all fitted models.}
\item{models}{List of fitted models.}
}
\description{
Algebraic difference scores as dependent variables in regression analysis conducted via structural equation modeling
}
\examples{
set.seed(342356)
d <- data.frame(
  var1 = rnorm(50),
  var2 = rnorm(50),
  x = rnorm(50)
)

dds <- diff_dv_sem(data = d, var1 = "var1", var2 = "var2", predictor = "x")
round(dds$aic, 2)
lavaan::summary(dds$models$opposite.ev,
  standardized = TRUE
)
}
\references{
Edwards, J. R. (1995). Alternatives to Difference Scores as Dependent Variables in the Study of Congruence in Organizational Research. Organizational Behavior and Human Decision Processes, 64(3), 307–324. \url{doi:10.1006/obhd.1995.1108}.
}
