% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_score_correlation.R
\name{diff_score_correlation}
\alias{diff_score_correlation}
\title{Test assumptions and infer about associations with algebraic difference scores}
\usage{
diff_score_correlation(
  data,
  var1,
  var2,
  center = FALSE,
  scale = FALSE,
  predictor,
  estimator = "MLR",
  level = 0.95,
  bound_l = -0.05,
  bound_u = 0.05,
  sampling.weights = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{var1}{Character string. Variable name of first component of difference score (Y_1).}

\item{var2}{Character string. Variable name of second component of difference score (Y_2).}

\item{center}{Logical. Are var1 and var2 centered around their grand mean? (Default FALSE)}

\item{scale}{Logical. Are var1 and var2 scaled with their pooled sd? (Default FALSE)}

\item{predictor}{Character string. Variable name of independent variable predicting difference.}

\item{estimator}{Character string. Estimator used in SEM (Default "MLR").}

\item{level}{Numeric. The confidence level required for the result output (Default .95)}

\item{bound_l}{Equivalence test lower bound based on SESOI (smallest effect size of interest; Default = -0.05).}

\item{bound_u}{Equivalence test upper bound based on SESOI (smallest effect size of interest; Default = 0.05).}

\item{sampling.weights}{Character string. Name of sampling weights variable.}
}
\value{
\item{descriptives}{Means, standard deviations, and intercorrelations.}
\item{parameter_estimates}{Parameter estimates from the structural equation model.}
\item{variances}{F test and Fligner-Killeen test for homogeneity of variances between var1 and var2.}
\item{transformed_data}{Data frame with variables used in SEM.}
\item{coef_sum_equivalence}{Equivalence test for sum of regression coefficiens for var1 and var2.}
\item{abs_test_one_sided}{One sided abs_test for positivity of abs(b_11-b_21)-abs(b_11+b_21).}
\item{results}{Summary of key results.}
}
\description{
Test assumptions and infer about associations with algebraic difference scores
}
\examples{
set.seed(342356)
d <- data.frame(
  var1 = rnorm(50),
  var2 = rnorm(50),
  x = rnorm(50)
)
diff_score_correlation(
  data = d, var1 = "var1", var2 = "var2",
  predictor = "x", center = TRUE, scale = TRUE
)
}
\references{
Edwards, J. R. (1995). Alternatives to Difference Scores as Dependent Variables in the Study of Congruence in Organizational Research. Organizational Behavior and Human Decision Processes, 64(3), 307–324. \url{doi:10.1006/obhd.1995.1108}.
}
