% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{example_big5}
\alias{example_big5}
\title{Example BIG5 personality dataset (random subset)}
\format{
A data frame with 200 rows and the same variables as the full BIG5 dataset.
\describe{
\item{gender}{Gender of respondent (1 = male, 2 = female)}
\item{country}{Country of respondent (here always "US")}
\item{age}{Age of respondent}
\item{N1-C10}{Personality items}

}
}
\source{
\url{http://openpsychometrics.org/_rawdata/}
}
\usage{
example_big5
}
\description{
A small subset of the BIG5 dataset from OpenPsychometrics, included as a
fallback for vignettes. This dataset consists of 200 U.S. respondents:
100 men and 100 women.
}
\keyword{datasets}
