% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{multid-deprecated}
\alias{multid-deprecated}
\title{Deprecated functions in multid}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item \code{\link{sem_dadas}}: This function is deprecated, and will
be removed in the next version of this package.
}
\itemize{
\item \code{\link{ml_dadas}}: This function is deprecated, and will
be removed in the next version of this package.
}
}
