% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{example_big5}
\alias{example_big5}
\title{Example BIG5 personality dataset (random subset)}
\format{
A data frame with 200 rows and the same variables as the full BIG5 dataset.
\describe{
\item{gender}{Gender of respondent (1 = male, 2 = female)}
\item{country}{Country of respondent (here always "US")}
\item{age}{Age of respondent}
\item{A1}{Agreeableness item 1}
\item{A2}{Agreeableness item 2}
\item{A3}{Agreeableness item 3}
\item{A4}{Agreeableness item 4}
\item{A5}{Agreeableness item 5}
\item{A6}{Agreeableness item 6}
\item{A7}{Agreeableness item 7}
\item{A8}{Agreeableness item 8}
\item{A9}{Agreeableness item 9}
\item{A10}{Agreeableness item 10}
\item{C1}{Conscientiousness item 1}
\item{C2}{Conscientiousness item 2}
\item{C3}{Conscientiousness item 3}
\item{C4}{Conscientiousness item 4}
\item{C5}{Conscientiousness item 5}
\item{C6}{Conscientiousness item 6}
\item{C7}{Conscientiousness item 7}
\item{C8}{Conscientiousness item 8}
\item{C9}{Conscientiousness item 9}
\item{C10}{Conscientiousness item 10}
\item{E1}{Extraversion item 1}
\item{E2}{Extraversion item 2}
\item{E3}{Extraversion item 3}
\item{E4}{Extraversion item 4}
\item{E5}{Extraversion item 5}
\item{E6}{Extraversion item 6}
\item{E7}{Extraversion item 7}
\item{E8}{Extraversion item 8}
\item{E9}{Extraversion item 9}
\item{E10}{Extraversion item 10}
\item{N1}{Neuroticism item 1}
\item{N2}{Neuroticism item 2}
\item{N3}{Neuroticism item 3}
\item{N4}{Neuroticism item 4}
\item{N5}{Neuroticism item 5}
\item{N6}{Neuroticism item 6}
\item{N7}{Neuroticism item 7}
\item{N8}{Neuroticism item 8}
\item{N9}{Neuroticism item 9}
\item{N10}{Neuroticism item 10}
\item{O1}{Openness item 1}
\item{O2}{Openness item 2}
\item{O3}{Openness item 3}
\item{O4}{Openness item 4}
\item{O5}{Openness item 5}
\item{O6}{Openness item 6}
\item{O7}{Openness item 7}
\item{O8}{Openness item 8}
\item{O9}{Openness item 9}
\item{O10}{Openness item 10}

}
}
\source{
\url{http://openpsychometrics.org/_rawdata/}
}
\usage{
example_big5
}
\description{
A small subset of the BIG5 dataset from OpenPsychometrics, included as a
fallback for vignettes. This dataset consists of 200 U.S. respondents:
100 men and 100 women.
}
\keyword{datasets}
