% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_dadas.R
\name{ml_dadas}
\alias{ml_dadas}
\title{Predicting algebraic difference scores in multilevel model}
\usage{
ml_dadas(model, predictor, diff_var, diff_var_values)
}
\arguments{
\item{model}{Multilevel model fitted with lmerTest.}

\item{predictor}{Character string. Variable name of independent variable predicting difference score.}

\item{diff_var}{Character string. A variable indicative of difference score components (two groups).}

\item{diff_var_values}{Vector. Values of the component score groups in diff_var.}
}
\value{
A data frame including regression coefficients for component scores and dadas (One sided dadas-test for positivity of abs(b_11-b_21)-abs(b_11+b_21)).
}
\description{
Predicting algebraic difference scores in multilevel model
}
